GridAliquotViewSelector: convert tests
